+++ app/Models/Zekering.php
@@ -6,7 +6,7 @@
  * Time: 22:20
  */
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
