+++ database/seeders/ZekeringenSeeder.php
@@ -14,7 +14,7 @@ class ZekeringenSeeder extends Seeder
     public function run()
     {
         //inserting test data
-        $zekering = new \App\Zekering();
+        $zekering = new \App\Models\Zekering();
         $zekering->text = "dit is een dummy tekst voor een zekering";
         $zekering->createdBy = 1;
         $zekering->score = 1;
@@ -24,7 +24,7 @@ class ZekeringenSeeder extends Seeder
         //add zekering(en)
         $zekering->save();
 
-        $zekering = new \App\Zekering();
+        $zekering = new \App\Models\Zekering();
         $zekering->text = "dit is een dummy tekst 2 voor een zekering";
         $zekering->createdBy = 1;
         $zekering->score = 1;
@@ -34,7 +34,7 @@ class ZekeringenSeeder extends Seeder
         //add zekering(en)
         $zekering->save();
 
-        $zekering = new \App\Zekering();
+        $zekering = new \App\Models\Zekering();
         $zekering->text = "dit is een dummy tekst 2 voor een zekering met een parent";
         $zekering->createdBy = 1;
         $zekering->score = 1;
